remove winning combination UC and implement in DO
