+++ bin/temp_game.rb
